requirements.txt
boto3==1.15.18
google-api-python-client==1.12.3
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
